test(diffusion): validation slice for the matrix-free apply-path
Add tests/ops/test_diffusion_apply.py -- the behavior-preservation gate,
comparing env.diffuse against the frozen pre-refactor dense baseline (not a
live recompute, so equivalence is vs an independent oracle):

- full-rank equivalence (rtol 1e-8, all modes, signed F, incl. polar/mesh);
- truncated equivalence in the M-weighted / dense-relative norm;
- linearity on signed fields (no positivity projection);
- compute_kernel action byte-identical to the baseline (rtol 1e-12);
- null-mode retention: rank >= n_components, exact mass conservation + H@1==1;
- grid-independence preserved; no leakage across a wall under truncation;
- W-component support gates (binned/resample): no spurious NaN, value matches
  dense where den >> tol, numerator not floored; diffusion_kde nonnegative;
- env.smooth negatives bounded M-weighted relative to dense (polar/mesh);
- single growable cache: grows on smaller sigma, sliced/reused, dropped on
  _state_version bump, not poisoned by compute_kernel, transient dense for
  near-full-rank;
- JAX backend: jax.Array return, in-JAX apply, jit + grad;
- test_perf_large_grid (slow): apply-path vs the dense baseline.

Also let JAX narrow the eigenbasis dtype at the boundary (no explicit float64
request), silencing the x64-disabled truncation warning.

Measured: 3600-bin grid dense expm 167 s / 763 MB -> env.diffuse 0.93 s / 61 MB
(181x faster, 13x lower peak memory); a 10k-bin grid smooths in ~15 s where the
dense expm is ~O(n^3) impractical. 75 non-slow tests pass; ruff + mypy clean.

+++ src/neurospatial/ops/diffusion.py
@@ -723,12 +723,16 @@ def diffusion_apply(
     if backend == "jax":
         import jax.numpy as jnp
 
+        # Cast the cached NumPy eigenbasis to the JAX default float (float64 when
+        # x64 is enabled, else float32 -- JAX narrows silently at the boundary, so
+        # no explicit dtype is requested here). The apply then runs entirely in
+        # jax.numpy, preserving jit / grad / GPU.
         return _apply_modes(
-            jnp.asarray(Q, dtype=jnp.float64),
-            jnp.asarray(Lam, dtype=jnp.float64),
-            jnp.asarray(volumes, dtype=jnp.float64),
+            jnp.asarray(Q),
+            jnp.asarray(Lam),
+            jnp.asarray(volumes),
             sigma,
-            jnp.asarray(fields, dtype=jnp.float64),
+            jnp.asarray(fields),
             mode,
             jnp,
         )
